orchestrator/requirements.txt
pyyaml
inquirer